+++ src/components/WorldMap/db/states.ts
@@ -6,6 +6,7 @@ import statesEuropaCzechia from './states/europe/czechia.json';
 import statesEuropaFrance from './states/europe/france.json';
 import statesEuropaGermany from './states/europe/germany.json';
 import statesEuropaPoland from './states/europe/poland.json';
+import statesEuropaSpain from './states/europe/spain.json';
 import statesEuropaUnitedKingdom from './states/europe/united-kingdom.json';
 
 /* Import types. */
@@ -63,6 +64,7 @@ export const getStates = (): TypeState[] => {
         ...(statesEuropaFrance.data as TypeState[]),
         ...(statesEuropaGermany.data as TypeState[]),
         ...(statesEuropaPoland.data as TypeState[]),
+        ...(statesEuropaSpain.data as TypeState[]),
         ...(statesEuropaUnitedKingdom.data as TypeState[]),
 
         /* Oceania states. */